requirements.txt
Flask==2.3.3
gunicorn==22.0.0
requests==2.32.3
beautifulsoup4==4.12.3
psycopg2-binary==2.9.9
cachetools==5.5.0
cryptography==43.0.1
PyJWT==2.8.0
python-dotenv==1.0.1
aiohttp==3.9.5
pytest==8.3.4
pytest-cov==6.0.0
